# backend/internal/api/docs/configer_swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Write-back-native GitOps configuration management. Configer discovers an existing repository's layout, exposes a parameter x instance grid over the repository's OWN files, and surgically edits them; every change flows to Git as commits, branches, pull requests, and merges. `.configer/` holds metadata only.\n\n## Repository scoping\nEvery resource endpoint exists in two shapes: the unscoped `/api/...` form acts on the default (first-connected) repository for single-repo deployments and older clients, and the identical routes are available per repository under `/api/repos/{repoId}/...`. Only the unscoped forms are documented here; to target a specific repository prefix any path with `/api/repos/{repoId}`.\n\n## Errors\nEvery 4xx/5xx response uses one envelope: a stable machine-readable `code`, a human-readable `error`, a `requestId` (also returned as the `X-Request-ID` header) that correlates to server logs, and - for validation failures - a `fields` array. Branch on `code`, never on the message text.\n\n## Correlation\nSend `X-Request-ID` to have it echoed and used as the correlation id; otherwise the server generates one. It is returned on every response.",
        "title": "Configer API",
...
        "contact": {
            "name": "Configer",
...
        },
...
    },
    "basePath": "/",
    "paths": {
...
        "/api/values": {
            "put": {
                "security": [
                    {
                        "CookieSession": []
                    }
                ],
                "description": "Stage a validated cell edit into the draft change request. The value is coerced to the parameter's declared type and validated; nothing touches Git until the draft is submitted. `action` defaults to \"set\"; \"reset\" drops the instance override; \"exclude\" removes the key from the instance's files. `scope:\"global\"` edits the shared value for every instance.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Editing \u0026 change requests"
                ],
                "summary": "Stage a value edit",
                "parameters": [
                    {
                        "description": "The edit",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ValueEditRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ValueStagedResponse"
                        }
                    },
                    "400": {
                        "description": "Malformed body or unsupported scope",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "404": {
                        "description": "Parameter or instance not found",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "422": {
                        "description": "Value failed coercion or validation",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "CookieSession": []
                    }
                ],
                "description": "Drop one pending edit (identified by paramId + instance) from the current draft.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Editing \u0026 change requests"
                ],
                "summary": "Revert a pending value edit",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Parameter id",
                        "name": "paramId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Instance name (empty for a global edit)",
                        "name": "instance",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.OKResponse"
                        }
                    },
                    "404": {
                        "description": "No draft",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    }
                }
            }
        },
        "/api/values/bulk": {
            "put": {
                "security": [
                    {
                        "CookieSession": []
                    }
                ],
                "description": "Stage the same parameter's edit on several instances at once. `edits` is a list of `{instance, value}`; `action` defaults to \"set\" (\"reset\"/\"exclude\" drop the override and ignore value). Each value is coerced and validated independently; invalid targets are reported in `results` while valid ones still stage. Nothing touches Git until the draft is submitted.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Editing \u0026 change requests"
                ],
                "summary": "Stage a value edit across many instances",
                "parameters": [
                    {
                        "description": "paramId, action, edits[]",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Malformed body",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "404": {
                        "description": "Parameter not found",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code is the stable, machine-readable classifier a client branches on.",
                    "type": "string",
                    "example": "not_found"
                },
                "error": {
                    "description": "Error is a human-readable, plain-language message safe to show a user.\nIt never contains git jargon, stack traces, credentials, or internal\nhostnames.",
                    "type": "string",
                    "example": "parameter not found"
                },
                "fields": {
                    "description": "Fields carries per-field validation failures (422 only).",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.FieldError"
                    }
                },
                "requestId": {
                    "description": "RequestID correlates this response with the access log and the\nX-Request-ID header, so an operator can find the exact request.",
                    "type": "string",
                    "example": "3f9a1c2b7e5d0a84"
                }
            }
        },
...
        "api.FieldError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string",
                    "example": "port"
                },
                "message": {
                    "type": "string",
                    "example": "must be between 1 and 65535"
                }
            }
        },
...
        "api.OKResponse": {
            "type": "object",
            "properties": {
                "ok": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "api.ValueEditRequest": {
            "type": "object",
            "properties": {
                "action": {
                    "description": "Action defaults to \"set\". \"reset\" drops the instance override; \"exclude\"\nremoves the key from the instance's files entirely.",
                    "type": "string",
                    "enum": [
                        "set",
                        "reset",
                        "exclude"
                    ],
                    "example": "set"
                },
                "instance": {
                    "type": "string",
                    "example": "prod-us-east"
                },
                "paramId": {
                    "type": "string",
                    "example": "net-admin-port"
                },
                "scope": {
                    "description": "Scope \"global\" edits the shared value for every instance; omit for a\nper-instance override.",
                    "type": "string",
                    "enum": [
                        "global"
                    ],
                    "example": ""
                },
                "value": {
                    "description": "Value is a type-appropriate value, coerced and validated server-side.",
                    "type": "string",
                    "example": "8443"
                }
            }
        },
        "api.ValueStagedResponse": {
            "type": "object",
            "properties": {
                "changeId": {
                    "type": "integer",
                    "example": 7
                },
                "ok": {
                    "type": "boolean",
                    "example": true
                },
                "pending": {
                    "type": "integer",
                    "example": 3
                },
                "value": {}
            }
        },
...
    },
    "securityDefinitions": {
        "CookieSession": {
            "description": "Session cookie set by the GitHub OAuth login flow (`GET /api/auth/login` -\u003e callback). Only required when the deployment has OAuth configured; single-user deployments need no authentication.",
            "type": "apiKey",
            "name": "configer_session",
            "in": "cookie"
        }
    }
}